requirements.txt
numpy
scikit-learn
pyspark
pandas
PyArrow
dataclasses
matplotlib
psycopg2 
python-dotenv 
tqdm
-e .
